Create Adjacency Matrix Count of Connected Component in undirected graph  depth first search.CPP
The time complexity of counting the number of connected components in an undirected graph using an adjacency matrix and DFS is O(N^2), where N is the number of vertices in the graph. This is because we need to traverse the entire adjacency matrix to initialize it, and in the worst case, we will visit all vertices and edges during the DFS traversal.

The space complexity of this algorithm is also O(N^2), since we need to store the adjacency matrix and the visited array, both of which have size N^2. Additionally, the stack space used by the DFS recursion can also be O(N) in the worst case, if the graph is a long linear chain.

However, if the graph is sparse (i.e., has relatively few edges compared to vertices), we can use an adjacency list instead of an adjacency matrix to reduce the space complexity to O(N + M), where M is the number of edges in the graph. The time complexity of the algorithm remains the same at O(N + M), since we still need to traverse all edges in the worst case.